+++ docassemble_webapp/docassemble/webapp/backend.py
@@ -32,6 +32,8 @@ from flask_wtf.csrf import generate_csrf
 from flask_login import current_user
 import docassemble.webapp.worker
 from docassemble.webapp.mailgun_mail import Mail as MailgunMail
+from bs4 import BeautifulSoup
+
 #sys.stderr.write("I am in backend\n")
 
 import docassemble.webapp.setup